backend/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
opencv-python-headless
google-generativeai
gTTS
pytest
httpx
google-cloud-tasks
# Pinned versions for stability from runbook
moviepy==1.0.3
imageio==2.31.2
imageio-ffmpeg==0.4.8
gunicorn==20.1.0
google-cloud-firestore
firebase-admin
pydantic-settings